Move lib/payment_gateways to app/lib/payment_gateways
This commit is one of some to move mostly of our lib/* to app/lib
in order to have eager loading on boot

This one moves lib/payment_gateways
